move notifications/type to common
