+++ packages/pn-pa-webapp/src/pages/components/NotificationDetail/DetailDocuments.tsx
@@ -2,7 +2,7 @@ import { Fragment, useEffect } from 'react';
 import { Grid, Typography, Button } from '@mui/material';
 import AttachFileIcon from '@mui/icons-material/AttachFile';
 
-import { NotificationDetail } from '../../../redux/notification/types';
+import { NotificationDetail } from '@pagopa-pn/pn-commons/src/types/Notifications';
 import { useAppDispatch, useAppSelector } from '../../../redux/hooks';
 import { getSentNotificationDocument } from '../../../redux/notification/actions';
 import { RootState } from '../../../redux/store';
