+++ packages/pn-pa-webapp/src/pages/components/NotificationDetail/DetailTimeline.tsx
@@ -12,7 +12,7 @@ import { Typography, Box, Button, Chip, Grid } from '@mui/material';
 import AttachFileIcon from '@mui/icons-material/AttachFile';
 import { getTime, getMonthString, getDay } from '@pagopa-pn/pn-commons';
 
-import { LegalFactId, NotificationDetail } from '../../../redux/notification/types';
+import { LegalFactId, NotificationDetail } from '@pagopa-pn/pn-commons/src/types/Notifications';
 import { getNotificationStatusLabelAndColorFromTimelineCategory } from '../../../utils/status.utility';
 import { useAppDispatch, useAppSelector } from '../../../redux/hooks';
 import { RootState } from '../../../redux/store';
