+++ packages/pn-pa-webapp/src/redux/notification/__test__/reducers.test.ts
@@ -2,7 +2,7 @@ import { NotificationsApi } from '../../../api/notifications/Notifications.api';
 import { mockAuthentication } from '../../auth/__test__/reducers.test';
 import { store } from '../../store';
 import { getSentNotification, getSentNotificationDocument, getSentNotificationLegalfact } from '../actions';
-import { LegalFactType, NotificationDetail } from '../types';
+import { LegalFactType, NotificationDetail } from '@pagopa-pn/pn-commons/src/types/Notifications';
 import { notificationToFe } from './test-utils';
 
 describe('Notification detail redux state tests', () => {
