+++ packages/pn-pa-webapp/src/utils/status.utility.ts
@@ -1,5 +1,5 @@
 import { getNotificationStatusLabelAndColor, NotificationStatus } from '@pagopa-pn/pn-commons';
-import { NotificationDetailTimeline, NotificationStatusHistory } from '../redux/notification/types';
+import { NotificationDetailTimeline, NotificationStatusHistory } from '@pagopa-pn/pn-commons/src/types/Notifications';
 
 export function getNotificationStatusLabelAndColorFromTimelineCategory(
   timelineStep: NotificationDetailTimeline,
