+++ packages/pn-personafisica-webapp/src/redux/notification/actions.ts
@@ -1,7 +1,7 @@
+import { LegalFactId, NotificationDetail } from '@pagopa-pn/pn-commons/src/types/Notifications';
 import { createAsyncThunk } from '@reduxjs/toolkit';
 
 import { NotificationsApi } from '../../api/notifications/Notifications.api';
-import { NotificationDetail, LegalFactId } from './types';
 
 export const getSentNotification = createAsyncThunk<NotificationDetail, string>(
   'getSentNotification',
